add ts-node, use ts in generation script
