+++ package.json
@@ -12,7 +12,7 @@
     "test:watch": "railway run --service bf8fe035-6790-4d5a-845a-1f5da75383a0 jest --watch",
     "test:coverage": "railway run --service bf8fe035-6790-4d5a-845a-1f5da75383a0 jest --coverage",
     "analyze": "railway run --service bf8fe035-6790-4d5a-845a-1f5da75383a0 cross-env ANALYZE=true next build",
-    "generate": "tsc src/utils.ts && railway run --service bf8fe035-6790-4d5a-845a-1f5da75383a0 node scripts/generateCharades.mjs",
+    "generate": "railway run --service bf8fe035-6790-4d5a-845a-1f5da75383a0 npx ts-node scripts/generateCharades.ts",
     "openai:play": "railway run --service bf8fe035-6790-4d5a-845a-1f5da75383a0 node scripts/testOpenAI.mjs",
     "cleanup": "./scripts/cleanup.sh",
     "ts": "tsc --noEmit --incremental",
@@ -31,6 +31,7 @@
     "react-dom": "18.2.0",
     "realm": "^11.10.2",
     "sharp": "^0.32.6",
+    "ts-node": "^10.9.2",
     "typescript": "^5.2.2",
     "uuid": "^9.0.0"
   },
@@ -46,6 +47,7 @@
     "@types/gtag.js": "^0.0.13",
     "@types/node": "^20.5.7",
     "@types/react-copy-to-clipboard": "^5.0.4",
+    "@types/uuid": "^9.0.7",
     "autoprefixer": "^10.4.7",
     "cross-env": "^7.0.3",
     "eslint": "8.20.0",
